Introduce basic raw scanner
This is the heart of the detection system, scanning blocks of data
for needles; things that would appear in a bitcoin wallet.

This is actually good enough as-is to find wallets and partial wallets,
but is missing two key additions; the ability to scan inside archives
(eg. looking for old backups and scanning inside them), and the ability
to handle encrypted file systems.